requirements.txt
absl-py==2.1.0
aiomysql==0.2.0
annotated-types==0.7.0
anyio==4.6.2.post1
asttokens==2.4.1
astunparse==1.6.3
Brotli @ file:///C:/b/abs_3d36mno480/croot/brotli-split_1714483178642/work
certifi @ file:///C:/b/abs_1fw_exq1si/croot/certifi_1725551736618/work/certifi
cffi==1.17.1
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.7
colorama @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/colorama_1699472650914/work
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
databases==0.9.0
debugpy==1.8.8
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
executing==2.1.0
fastapi==0.115.4
filelock @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/filelock_1701807523603/work
flatbuffers==24.3.25
fonttools==4.55.0
fsspec==2024.10.0
gast==0.6.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.68.1
h11==0.14.0
h5py==3.12.1
huggingface-hub==0.26.3
idna==3.10
imageio==2.36.1
ipykernel @ file:///C:/b/abs_5etv6sld_j/croot/ipykernel_1728665608850/work
ipython==8.29.0
jedi @ file:///C:/b/abs_1b8kmj7rrm/croot/jedi_1721058359741/work
Jinja2 @ file:///C:/b/abs_ddyeze2h6_/croot/jinja2_1730903330414/work
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core @ file:///C:/b/abs_beftpbuevw/croot/jupyter_core_1718818307097/work
jwt==1.3.1
keras==3.5.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
Mako==1.3.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft @ file:///C:/Users/dev-admin/mkl/mkl_fft_1730823082242/work
mkl_random @ file:///C:/Users/dev-admin/mkl/mkl_random_1730822522280/work
ml-dtypes==0.4.1
mpmath @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/mpmath_1699484863771/work
mysql==0.0.3
mysql-connector-python==9.1.0
mysqlclient==2.2.6
namex==0.0.8
nest-asyncio @ file:///C:/b/abs_65d6lblmoi/croot/nest-asyncio_1708532721305/work
networkx @ file:///C:/b/abs_3bxnu56g9d/croot/networkx_1717597507456/work
numpy @ file:///C:/b/abs_0123vcxhf8/croot/numpy_and_numpy_base_1725470331966/work/dist/numpy-2.0.1-cp312-cp312-win_amd64.whl#sha256=e3e9d7fcaa32e9aa86aa46a7191a47fc4c4b35942462bbbedd17c45ef934dccd
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow @ file:///C:/b/abs_2cm0qc1vxk/croot/pillow_1731594717982/work
platformdirs==4.3.6
prompt_toolkit==3.0.48
protobuf==5.29.0
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pycuda==2024.1.2
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==1.7.1
PyMySQL==1.1.1
pyparsing==3.2.0
PySocks @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/pysocks_1699473336188/work
python-dateutil @ file:///C:/b/abs_3au_koqnbs/croot/python-dateutil_1716495777160/work
python-dotenv==1.0.1
python-multipart==0.0.17
pytools==2024.1.21
pytz==2024.2
pywin32==305.1
PyYAML @ file:///C:/b/abs_14xkfs39bx/croot/pyyaml_1728657968772/work
pyzmq==26.2.0
regex==2024.11.6
requests @ file:///C:/b/abs_c3508vg8ez/croot/requests_1731000584867/work
rich==13.9.4
safetensors==0.4.5
scikit-image==0.25.0
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.1.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.1
SQLAlchemy==2.0.36
stack-data==0.6.3
starlette==0.41.2
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.12.12
tokenizers==0.21.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado @ file:///C:/b/abs_7bua0304mj/croot/tornado_1718740122405/work
tqdm==4.67.1
traitlets @ file:///C:/b/abs_bfsnoxl4pq/croot/traitlets_1718227069245/work
transformers==4.47.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3 @ file:///C:/b/abs_9a_f8h_bn2/croot/urllib3_1727769836930/work
uvicorn==0.32.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.44.0
win-inet-pton @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/win_inet_pton_1699472992992/work
wrapt==1.17.0
